chore: move dataclasses
